Improve search filter functionality
With this change, the search filter included in the archives.html
template defaults to exclusively the category clicked instead of a
combination with previously selected category and offers a way to
multi-select by holding down the CTRL key if multiple selection is
required. Since the CTRL key behaviour is pretty standard across
UIs, *no tips or manual is provided, as it is expected behaviour
with selections.

*NOTE:
  - a tip can be added
  - a UI element indicating the current selection mode can be added
    to visually show clearly what is the current selection mode