ft: Update timestamp insertion across the board
Include year in YYYY format on all post timestamps,
as some listings are not grouped by Year and it is
nice to see the year of post along-side the date on
those listings.

This commit also includes the css for styling those
newly formatted timestamps.

This also implements the same change in the JS based
listings rendered by search.ts on the archives pages
using the updated index.json